2025/10/22	3347. Maximum Frequency of an Element After Performing Operations II

+++ daily/3347.maximum-frequency-of-an-element-after-performing-operations-ii.cpp
@@ -6,7 +6,7 @@ class Solution {
     public:
         int maxFrequency(vector<int>& nums, int k, int numOperations) {
             int mx = *max_element(nums.begin(), nums.end());
-            map<long long, int> count, df;
+            map<int, int> count, df;
             for(auto x: nums) {
                 count[x]++;
                 df[x]++;
